add user restriction functionality to pages

+++ base/views.py
@@ -13,6 +13,9 @@ from .forms import RoomForm
 
 
 def loginPage(request: HttpResponse):
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -66,6 +69,7 @@ def room(request: HttpRequest, pk: str) -> HttpResponse:
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request: HttpRequest):
     form = RoomForm()
 
@@ -79,9 +83,13 @@ def createRoom(request: HttpRequest):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request: HttpRequest, pk: str) -> HttpResponse:
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not the author of this room. >:(')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -93,8 +101,13 @@ def updateRoom(request: HttpRequest, pk: str) -> HttpResponse:
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request: HttpRequest, pk: str) -> HttpResponse:
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not the author of this room. >:(')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
